requirements.txt
boto3==1.9.237
botocore==1.12.237
Django==2.2.9
django-crispy-forms==1.7.2
django-storages==1.7.2
docutils==0.15.2
flake8==3.7.8
icalendar==4.0.3
jmespath==0.9.4
Pillow==6.2.0
psycopg2-binary==2.8.3
pycryptodome==3.9.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
